# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HistoryExportResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "exportedAt": {
            "title": "Exportedat",
            "type": "string"
          },
          "filters": {
            "$ref": "#/components/schemas/HistoryExportFilters"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/HistoryRecord"
            },
            "title": "Records",
            "type": "array"
          }
        },
        "required": [
          "exportedAt",
          "filters",
          "count",
          "records"
        ],
        "title": "HistoryExportResponse",
        "type": "object"
      },
      "HistoryListResponse": {
        "properties": {
          "records": {
            "items": {
              "$ref": "#/components/schemas/HistoryRecord"
            },
            "title": "Records",
            "type": "array"
          }
        },
        "required": [
          "records"
        ],
        "title": "HistoryListResponse",
        "type": "object"
      },
...
      "ImageRecognitionRequest": {
        "properties": {
          "fileName": {
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "imageDataUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imagedataurl"
          },
          "sizeLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sizelabel"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          }
        },
        "required": [
          "fileName"
        ],
        "title": "ImageRecognitionRequest",
        "type": "object"
      },
      "ImageRecognitionResponse": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "historyRecord": {
            "$ref": "#/components/schemas/HistoryRecord"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "probabilities": {
            "items": {
              "$ref": "#/components/schemas/ProbabilityItem"
            },
            "title": "Probabilities",
            "type": "array"
          }
        },
        "required": [
          "label",
          "confidence",
          "explanation",
          "probabilities",
          "historyRecord"
        ],
        "title": "ImageRecognitionResponse",
        "type": "object"
      },
...
      "MuseumVisionRequest": {
        "properties": {
          "dimensions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dimensions"
          },
          "fileName": {
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format"
          },
          "imageDataUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imagedataurl"
          },
          "sizeLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sizelabel"
          },
          "sourceMode": {
            "default": "sample",
            "enum": [
              "sample",
              "upload"
            ],
            "title": "Sourcemode",
            "type": "string"
          }
        },
        "required": [
          "fileName"
        ],
        "title": "MuseumVisionRequest",
        "type": "object"
      },
      "MuseumVisionResponse": {
        "properties": {
          "artworkClue": {
            "$ref": "#/components/schemas/MuseumArtworkClue"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "dimensions": {
            "title": "Dimensions",
            "type": "string"
          },
          "format": {
            "title": "Format",
            "type": "string"
          },
          "historyRecord": {
            "$ref": "#/components/schemas/HistoryRecord"
          },
          "institution": {
            "title": "Institution",
            "type": "string"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/MuseumMatch"
            },
            "title": "Matches",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sizeLabel": {
            "title": "Sizelabel",
            "type": "string"
          },
          "sourceNote": {
            "title": "Sourcenote",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "uploadedAt": {
            "title": "Uploadedat",
            "type": "string"
          }
        },
        "required": [
          "name",
          "format",
          "dimensions",
          "sizeLabel",
          "sourceNote",
          "uploadedAt",
          "institution",
          "confidence",
          "description",
          "artworkClue",
          "tags",
          "matches",
          "historyRecord"
        ],
        "title": "MuseumVisionResponse",
        "type": "object"
      },
...
      "MuseumVisionTagExportRequest": {
        "properties": {
          "fileName": {
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "institution": {
            "minLength": 1,
            "title": "Institution",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "fileName",
          "institution",
          "tags"
        ],
        "title": "MuseumVisionTagExportRequest",
        "type": "object"
      },
      "ProbabilityItem": {
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "label",
          "value"
        ],
        "title": "ProbabilityItem",
        "type": "object"
      },
...
      "SearchResponse": {
        "properties": {
          "keyword": {
            "title": "Keyword",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchSuggestion"
            },
            "title": "Results",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "keyword",
          "total",
          "results"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
...
      "SentimentAnalysisRequest": {
        "properties": {
          "text": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "SentimentAnalysisRequest",
        "type": "object"
      },
      "SentimentAnalysisResponse": {
        "properties": {
          "completedAt": {
            "title": "Completedat",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "englishLabel": {
            "enum": [
              "Positive",
              "Neutral",
              "Negative"
            ],
            "title": "Englishlabel",
            "type": "string"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "historyRecord": {
            "$ref": "#/components/schemas/HistoryRecord"
          },
          "label": {
            "enum": [
              "正面",
              "中性",
              "负面"
            ],
            "title": "Label",
            "type": "string"
          },
          "negativeMatches": {
            "items": {
              "$ref": "#/components/schemas/KeywordMatch"
            },
            "title": "Negativematches",
            "type": "array"
          },
          "positiveMatches": {
            "items": {
              "$ref": "#/components/schemas/KeywordMatch"
            },
            "title": "Positivematches",
            "type": "array"
          },
          "processingTime": {
            "title": "Processingtime",
            "type": "string"
          },
          "providerStatusMessage": {
            "title": "Providerstatusmessage",
            "type": "string"
          },
          "providerUsed": {
            "enum": [
              "local",
              "deepseek"
            ],
            "title": "Providerused",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "status": {
            "enum": [
              "已完成",
              "待分析"
            ],
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "taskId": {
            "title": "Taskid",
            "type": "string"
          },
          "usedFallback": {
            "title": "Usedfallback",
            "type": "boolean"
          }
        },
        "required": [
          "label",
          "englishLabel",
          "confidence",
          "score",
          "tags",
          "positiveMatches",
          "negativeMatches",
          "explanation",
          "status",
          "processingTime",
          "taskId",
          "completedAt",
          "providerUsed",
          "usedFallback",
          "providerStatusMessage",
          "historyRecord"
        ],
        "title": "SentimentAnalysisResponse",
        "type": "object"
      },
...
      "TextGenerationRequest": {
        "properties": {
          "quantity": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Quantity",
            "type": "integer"
          },
          "theme": {
            "maxLength": 80,
            "title": "Theme",
            "type": "string"
          },
          "tone": {
            "enum": [
              "正式",
              "活泼",
              "科技感",
              "文艺"
            ],
            "title": "Tone",
            "type": "string"
          },
          "type": {
            "enum": [
              "标题",
              "宣传语",
              "短文案",
              "诗意表达"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "theme",
          "tone",
          "type",
          "quantity"
        ],
        "title": "TextGenerationRequest",
        "type": "object"
      },
      "TextGenerationResponse": {
        "properties": {
          "generatedAt": {
            "title": "Generatedat",
            "type": "string"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/GenerationHistoryItem"
            },
            "title": "History",
            "type": "array"
          },
          "historyEntry": {
            "$ref": "#/components/schemas/GenerationHistoryItem"
          },
          "historyRecord": {
            "$ref": "#/components/schemas/HistoryRecord"
          },
          "outputs": {
            "items": {
              "$ref": "#/components/schemas/GenerationOutput"
            },
            "title": "Outputs",
            "type": "array"
          },
          "providerStatusMessage": {
            "title": "Providerstatusmessage",
            "type": "string"
          },
          "providerUsed": {
            "enum": [
              "local",
              "deepseek"
            ],
            "title": "Providerused",
            "type": "string"
          },
          "qualityMetrics": {
            "items": {
              "$ref": "#/components/schemas/GenerationQualityMetric"
            },
            "title": "Qualitymetrics",
            "type": "array"
          },
          "qualityTip": {
            "title": "Qualitytip",
            "type": "string"
          },
          "toneKeywords": {
            "items": {
              "type": "string"
            },
            "title": "Tonekeywords",
            "type": "array"
          },
          "usedFallback": {
            "title": "Usedfallback",
            "type": "boolean"
          }
        },
        "required": [
          "outputs",
          "historyEntry",
          "history",
          "qualityMetrics",
          "qualityTip",
          "toneKeywords",
          "generatedAt",
          "providerUsed",
          "usedFallback",
          "providerStatusMessage",
          "historyRecord"
        ],
        "title": "TextGenerationResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "多模态 AI 课程成果平台 Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/history": {
      "get": {
        "operationId": "history_api_v1_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "title": "Keyword",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "module",
            "required": false,
            "schema": {
              "default": "全部",
              "title": "Module",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "全部",
              "title": "Status",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "History"
      }
    },
    "/api/v1/history/export": {
      "get": {
        "operationId": "export_history_api_v1_history_export_get",
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "title": "Keyword",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "module",
            "required": false,
            "schema": {
              "default": "全部",
              "title": "Module",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "全部",
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "enum": [
                "json",
                "csv"
              ],
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryExportResponse"
                }
              },
              "text/csv": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Exported history records as a downloadable JSON or CSV attachment.",
            "headers": {
              "Content-Disposition": {
                "description": "Suggested attachment filename for the exported history file.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export History"
      }
    },
...
    "/api/v1/image-recognition/predict": {
      "post": {
        "operationId": "image_recognition_api_v1_image_recognition_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageRecognitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageRecognitionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Image Recognition"
      }
    },
    "/api/v1/museum-vision/analyze": {
      "post": {
        "operationId": "museum_vision_api_v1_museum_vision_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MuseumVisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MuseumVisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Museum Vision"
      }
    },
    "/api/v1/museum-vision/export-tags": {
      "post": {
        "operationId": "museum_vision_export_tags_api_v1_museum_vision_export_tags_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MuseumVisionTagExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/plain": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Exported the current museum analysis tags as a downloadable text attachment.",
            "headers": {
              "Content-Disposition": {
                "description": "Suggested attachment filename for the exported museum tag file.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Museum Vision Export Tags"
      }
    },
...
    "/api/v1/search": {
      "get": {
        "operationId": "search_api_v1_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "title": "Keyword",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search"
      }
    },
    "/api/v1/sentiment-analysis/analyze": {
      "post": {
        "operationId": "sentiment_analysis_api_v1_sentiment_analysis_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SentimentAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SentimentAnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sentiment Analysis"
      }
    },
...
    "/api/v1/text-generation/generate": {
      "post": {
        "operationId": "text_generation_api_v1_text_generation_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextGenerationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Text Generation"
      }
    },
...
  }
}